// packages/core/src/squads/generated/types/period.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
  getEnumDecoder,
  getEnumEncoder,
} from "@solana/kit";

/** The reset period of the spending limit. */
export enum Period {
  OneTime,
  Day,
  Week,
  Month,
}

export type PeriodArgs = Period;

export function getPeriodEncoder(): FixedSizeEncoder<PeriodArgs> {
  return getEnumEncoder(Period);
}

export function getPeriodDecoder(): FixedSizeDecoder<Period> {
  return getEnumDecoder(Period);
}

export function getPeriodCodec(): FixedSizeCodec<PeriodArgs, Period> {
  return combineCodec(getPeriodEncoder(), getPeriodDecoder());
}